s/double/float

+++ db/migrate/20220308020701_add_spread_to_matches.rb
@@ -1,5 +1,5 @@
 class AddSpreadToMatches < ActiveRecord::Migration[6.0]
   def change
-    add_column :matches, :spread, :double
+    add_column :matches, :spread, :float
   end
 end
